requirements.txt
streamlit>=1.31.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
pandas>=2.2.0
